[fix] 위치권한 permission Splash -> Login으로 이전
